--- x/auth/keeper/params.go
@@ -1,21 +1,0 @@
-package keeper
-
-import (
-	"context"
-
-	"github.com/cosmos/cosmos-sdk/x/auth/types"
-)
-
-// SetParams sets the auth module's parameters.
-func (ak AccountKeeper) SetParams(ctx context.Context, params types.Params) error {
-	if err := params.Validate(); err != nil {
-		return err
-	}
-	return ak.ParamsState.Set(ctx, params)
-}
-
-// GetParams gets the auth module's parameters.
-func (ak AccountKeeper) GetParams(ctx context.Context) (params types.Params) {
-	params, _ = ak.ParamsState.Get(ctx)
-	return params
-}
